An application that shows whether there is a risk of heart disease based on the patient information entered, using machine learning models and data processing method.
Heart diseases lead to many serious diseases and even death today. Many people live their lives unaware that they are at risk of heart disease, but they do not realize this until the sad and frightening consequences happen to them. For this reason, early diagnosis saves lives and is important for the life of individuals. With this project I have developed to solve this problem, I aim to facilitate the early diagnosis of heart diseases by developing an interface that can be used by users, hospitals, and some health institutions, namely everyone. First, I found the most suitable machine learning algorithm for the data set. Then I designed a prediction system using this SVM algorithm. In order to make this system available to everyone, I developed an interface and designed a system that predicts whether they carry a risk of heart disease by asking certain questions to the users. . The dataset I use in the project is the Heart Disease dataset we got from the UCI Machine Learning Repository. Although there are 76 attributes in this dataset, all published studies only use a subset of 14 of them and dataset consist of 303 rows. The input of the project is the health data I receive from the user in line with the features in dataset. The output of the project is the result of either there is a risk of heart disease or there is no risk of heart disease, which will be calculated and returned according to the data entered by the user as a result of all operations. The machine learning model we chose to use is Linear SVM. While choosing the most suitable algorithm for the project, I considered the training period, the number of features, the output of the project, the number of columns and parameters in the dataset, and the inputs and output of our project. In this way, we tried many suitable algorithms and decided to use the Linear SVM (Support Vector Machine) algorithm with the highest performance and accuracy percentage. Using this algorithm, we achieved almost 86% accuracy. I found this algorithm suitable for our project, because SVM is a classification algorithm. It tries to find the best line called hyperplane separating the two classes. The algorithm ensures that the line to be drawn is set to pass from the farthest place to its elements in two classes. In my project there are 2 classes, those at risk of heart disease and those without. Linear SVM is used for linearly separable data like in my Project.